PoC for generating JSON-formatted logs.
NB: Probably not that efficient to write strings like this, but will ensure
that the encoded data contains valid JSON.

Iteration on this would probably be to write json-based log lines with real metadata, rather
than the "make a timestamp + message into a json log".

hopefully though, this will hold an entire stacktrace as one entity for now, so parsing becomes
much easier.